updated and moved several files around and created new files relating to notes
