+++ src/components/homeAndAbout/quickDash.tsx
@@ -4,9 +4,9 @@ import {
   readVisits,
   writeVisits,
   type Visit,
-} from '@/components/shared/hooks/useVisitTracker';
+} from '@/hooks/storage/useVisitTracker';
 import { resolveBreadcrumbTrail } from '@/domain/navigation/breadcrumbs';
-import { SITE_NAME } from '@/components/shared/dynamicSiteName';
+import { SITE_NAME } from '@/config/dynamicSiteName';
 
 /* ---------- Title cache ---------- */
 const TITLES_KEY = "page_titles_v1";
