--- src/components/pageComponents/notes/homeText/utils/themeHome.ts
@@ -1,7 +1,0 @@
-export const homeTheme = {
-    h1: 'home-h1',
-    h2: 'home-h2',
-    p: 'home-p',
-    link: 'home-link',
-    blockquote: 'home-quote',
-  };
