+++ src/components/shared/analytics/routeTracker.tsx
@@ -1,8 +1,8 @@
 import { useLocation } from "react-router-dom";
-import { useVisitTracker } from '@/components/shared/hooks/useVisitTracker';
+import { useVisitTracker } from '@/hooks/storage/useVisitTracker';
 
 export default function RouteTracker() {
   const { pathname } = useLocation();
-  useVisitTracker(pathname); // tracks EVERY page, not just Home
+  useVisitTracker(pathname);
   return null;
 }
