+++ src/lib/firebaseNotes/FirebaseNotesAdapter.ts
@@ -1,4 +1,4 @@
-import type { NotesAdapter, NoteMeta, NoteDoc } from '@/lib/notesAdapter';
+import type { NotesAdapter, NoteMeta, NoteDoc } from '@/types/notes/adapter';
 import { dbNotes, SITE_ID } from '@/firebase/client';
 import {
   collection, getDocs, query, where, orderBy, limit, FirestoreDataConverter, DocumentData
