--- src/lib/notes.ts
@@ -1,6 +1,0 @@
-import { FirebaseNotesAdapter } from "./FirebaseNotesAdapter";
-import type { NotesAdapter } from "./notesAdapter";
-
-const adapter: NotesAdapter = new FirebaseNotesAdapter();
-
-export default adapter;
