+++ src/types/notes/index.ts
@@ -1,0 +1,1 @@
+export * from '@/types/notes/adapter';
